fix: refactoring tests files
