+++ cmd/server/handler/index_test.go
@@ -2,9 +2,27 @@ package controllers_test
 
 import (
 	"bytes"
+	"fmt"
 	"net/http"
 	"net/http/httptest"
+
+	"github.com/gin-gonic/gin"
+	"github.com/marcoglnd/mercado-fresco-packmain/cmd/server/routes"
 )
+
+func getPathUrl(url string) string {
+	PATH := "/api/v1"
+	return fmt.Sprintf("%s%s", PATH, url)
+}
+
+func createServer() *gin.Engine {
+	gin.SetMode(gin.TestMode)
+	router := gin.Default()
+	routerGroup := router.Group(getPathUrl(""))
+	routes.AddRoutes(routerGroup)
+
+	return router
+}
 
 func createRequestTest(
 	method string,
